requirements.txt
absl-py==2.1.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
Werkzeug==3.1.3
